+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `Ellipsoid`: Added `getNorthUpEastFrame`.
+
 ## [0.3.0] - 2025-07-05
 
 ### Changed